refactor(dispatcher): remove ADispatcher. - useless abstraction - only keep CommandDispatcher - move CommandDispatcher to Structures root
